requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
apify-client==1.7.1
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.1
cryptography==41.0.8
mysql-connector-python==8.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
